Add dummy button for sign-in
