+++ client/src/app/app-routing.module.ts
@@ -1,17 +1,14 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
-// To switch between screens (login and tabs)
-// change paths: 'tabs' to ''
-// change redirectTo: '/login' to 'tabs'
 
 const routes: Routes = [
   {
-    path: 'tabs',
-    loadChildren: () => import('./tabs/tabs.module').then(m => m.TabsPageModule)
+    path: 'login',
+    loadChildren: () => import('./public/login/login.module').then( m => m.LoginPageModule),
   },
   {
-    path: 'login',
-    loadChildren: () => import('./login/login.module').then( m => m.LoginPageModule)
+    path: 'tabs',
+    loadChildren: () => import('./members/tabs/tabs.module').then(m => m.TabsPageModule),
   },
   {
     path: '',
