+++ client/src/app/members/search-menu/search-menu.page.scss
