+++ client/src/app/members/tabs/tabs-routing.module.ts
@@ -4,7 +4,7 @@ import { TabsPage } from './tabs.page';
 
 const routes: Routes = [
   {
-    path: 'tabs',
+    path: '',
     component: TabsPage,
     children: [
       {
@@ -35,11 +35,6 @@ const routes: Routes = [
       }
     ]
   },
-  {
-    path: '',
-    redirectTo: '/tabs/search',
-    pathMatch: 'full'
-  }
 ];
 
 @NgModule({
