+++ client/src/app/public/login/login.page.html
@@ -7,7 +7,7 @@
       <ion-button class="register-btn" routerLink="./registration" routerDirection="forward" routerLinkActive="active">
         Register
       </ion-button>
-      <ion-button class="sign-in-btn" type="submit" expand="block" shape="round">
+      <ion-button class="sign-in-btn" (click)="signIn()" expand='full' shape="round">
         Sign in
       </ion-button>
     </ion-buttons>
