+++ client/src/app/public/registration/registration-routing.module.ts
