+++ client/src/app/public/registration/registration.page.spec.ts
