+++ client/src/app/public/registration/registration.page.ts
